move dataloader into page
